[ethers-v5] refactor: move common file to static
Moves commons.ts inline code to a static directory outside src. This is
to follow same pattern followed in other targets as well. This also renames
commons.ts to common.d.ts as in most places we have "common" while
here we have "commons", looks good to have similar naming. Also since
it is outside src dir, the build fails if we have common.ts, so that's
why it's converted to '.d.ts' and import is changed to type import.
